move functions that interact with the Crucible agent to common_storage.rs
