+++ nexus/src/app/sagas/disk_create.rs
@@ -3,8 +3,8 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 use super::{
-    ensure_all_datasets_and_regions, ActionRegistry, NexusActionContext,
-    NexusSaga, SagaInitError, ACTION_GENERATE_ID,
+    common_storage::ensure_all_datasets_and_regions, ActionRegistry,
+    NexusActionContext, NexusSaga, SagaInitError, ACTION_GENERATE_ID,
 };
 use crate::app::sagas::NexusAction;
 use crate::context::OpContext;
