+++ nexus/src/app/sagas/mod.rs
@@ -9,19 +9,8 @@
 // correctness, idempotence, etc.  The more constrained this interface is, the
 // easier it will be to test, version, and update in deployed systems.
 
-use crate::db;
-use crate::db::identity::Asset;
 use crate::saga_interface::SagaContext;
-use anyhow::anyhow;
-use crucible_agent_client::{
-    types::{CreateRegion, RegionId, State as RegionState},
-    Client as CrucibleAgentClient,
-};
-use futures::StreamExt;
 use lazy_static::lazy_static;
-use omicron_common::api::external::Error;
-use omicron_common::backoff::{self, BackoffError};
-use slog::Logger;
 use std::sync::Arc;
 use steno::new_action_noop_undo;
 use steno::ActionContext;
@@ -36,6 +25,8 @@ pub mod instance_create;
 pub mod instance_migrate;
 pub mod snapshot_create;
 pub mod volume_delete;
+
+pub mod common_storage;
 
 #[derive(Debug)]
 pub struct NexusSagaType;
@@ -120,120 +111,4 @@ pub(super) async fn saga_generate_uuid<UserType: SagaType>(
     _: ActionContext<UserType>,
 ) -> Result<Uuid, ActionError> {
     Ok(Uuid::new_v4())
-}
-
-// Arbitrary limit on concurrency, for operations issued on multiple regions
-// within a disk at the same time.
-const MAX_CONCURRENT_REGION_REQUESTS: usize = 3;
-
-/// Call out to Crucible agent and perform region creation.
-pub async fn ensure_region_in_dataset(
-    log: &Logger,
-    dataset: &db::model::Dataset,
-    region: &db::model::Region,
-) -> Result<crucible_agent_client::types::Region, Error> {
-    let url = format!("http://{}", dataset.address());
-    let client = CrucibleAgentClient::new(&url);
-
-    let region_request = CreateRegion {
-        block_size: region.block_size().to_bytes(),
-        extent_count: region.extent_count().try_into().unwrap(),
-        extent_size: region.blocks_per_extent().try_into().unwrap(),
-        // TODO: Can we avoid casting from UUID to string?
-        // NOTE: This'll require updating the crucible agent client.
-        id: RegionId(region.id().to_string()),
-        encrypted: region.encrypted(),
-        cert_pem: None,
-        key_pem: None,
-        root_pem: None,
-    };
-
-    let create_region = || async {
-        let region = client
-            .region_create(&region_request)
-            .await
-            .map_err(|e| BackoffError::Permanent(e.into()))?;
-        match region.state {
-            RegionState::Requested => Err(BackoffError::transient(anyhow!(
-                "Region creation in progress"
-            ))),
-            RegionState::Created => Ok(region),
-            _ => Err(BackoffError::Permanent(anyhow!(
-                "Failed to create region, unexpected state: {:?}",
-                region.state
-            ))),
-        }
-    };
-
-    let log_create_failure = |_, delay| {
-        warn!(
-            log,
-            "Region requested, not yet created. Retrying in {:?}", delay
-        );
-    };
-
-    let region = backoff::retry_notify(
-        backoff::internal_service_policy(),
-        create_region,
-        log_create_failure,
-    )
-    .await
-    .map_err(|e| Error::internal_error(&e.to_string()))?;
-
-    Ok(region.into_inner())
-}
-
-pub async fn ensure_all_datasets_and_regions(
-    log: &Logger,
-    datasets_and_regions: Vec<(db::model::Dataset, db::model::Region)>,
-) -> Result<
-    Vec<(db::model::Dataset, crucible_agent_client::types::Region)>,
-    ActionError,
-> {
-    let request_count = datasets_and_regions.len();
-
-    // Allocate regions, and additionally return the dataset that the region was
-    // allocated in.
-    let datasets_and_regions: Vec<(
-        db::model::Dataset,
-        crucible_agent_client::types::Region,
-    )> = futures::stream::iter(datasets_and_regions)
-        .map(|(dataset, region)| async move {
-            match ensure_region_in_dataset(log, &dataset, &region).await {
-                Ok(result) => Ok((dataset, result)),
-                Err(e) => Err(e),
-            }
-        })
-        // Execute the allocation requests concurrently.
-        .buffer_unordered(std::cmp::min(
-            request_count,
-            MAX_CONCURRENT_REGION_REQUESTS,
-        ))
-        .collect::<Vec<
-            Result<
-                (db::model::Dataset, crucible_agent_client::types::Region),
-                Error,
-            >,
-        >>()
-        .await
-        .into_iter()
-        .collect::<Result<
-            Vec<(db::model::Dataset, crucible_agent_client::types::Region)>,
-            Error,
-        >>()
-        .map_err(ActionError::action_failed)?;
-
-    // Assert each region has the same block size, otherwise Volume creation
-    // will fail.
-    let all_region_have_same_block_size = datasets_and_regions
-        .windows(2)
-        .all(|w| w[0].1.block_size == w[1].1.block_size);
-
-    if !all_region_have_same_block_size {
-        return Err(ActionError::action_failed(Error::internal_error(
-            "volume creation will fail due to block size mismatch",
-        )));
-    }
-
-    Ok(datasets_and_regions)
 }
