+++ nexus/src/app/sagas/snapshot_create.rs
@@ -82,8 +82,8 @@
 //!
 
 use super::{
-    ensure_all_datasets_and_regions, ActionRegistry, NexusActionContext,
-    NexusSaga, SagaInitError, ACTION_GENERATE_ID,
+    common_storage::ensure_all_datasets_and_regions, ActionRegistry,
+    NexusActionContext, NexusSaga, SagaInitError, ACTION_GENERATE_ID,
 };
 use crate::app::sagas::NexusAction;
 use crate::context::OpContext;
